+++ applications/external/weather_station/protocols/auriol_ahfl.h
@@ -49,7 +49,7 @@ void ws_protocol_decoder_auriol_ahfl_feed(void* context, bool level, uint32_t du
  * @param context Pointer to a WSProtocolDecoderAuriol_AHFL instance
  * @return hash Hash sum
  */
-uint8_t ws_protocol_decoder_auriol_ahfl_get_hash_data(void* context);
+uint32_t ws_protocol_decoder_auriol_ahfl_get_hash_data(void* context);
 
 /**
  * Serialize data WSProtocolDecoderAuriol_AHFL.
